+++ crates/entrypoint/src/cli/command/azure/policy/assignment/mod.rs
@@ -1,10 +1,12 @@
 pub mod azure_policy_assignment;
 pub mod azure_policy_assignment_browse;
 pub mod azure_policy_assignment_list;
+pub mod azure_policy_assignment_show;
 
 pub use azure_policy_assignment::AzurePolicyAssignmentCommand;
 pub use azure_policy_assignment_browse::AzurePolicyAssignmentBrowseArgs;
 pub use azure_policy_assignment_list::AzurePolicyAssignmentListArgs;
+pub use azure_policy_assignment_show::AzurePolicyAssignmentShowArgs;
 use clap::Args;
 use eyre::Result;