+++ crates/entrypoint/src/cli/command/azure/policy/definition/mod.rs
@@ -1,10 +1,12 @@
 pub mod azure_policy_definition;
 pub mod azure_policy_definition_browse;
 pub mod azure_policy_definition_list;
+pub mod azure_policy_definition_show;
 
 pub use azure_policy_definition::AzurePolicyDefinitionCommand;
 pub use azure_policy_definition_browse::AzurePolicyDefinitionBrowseArgs;
 pub use azure_policy_definition_list::AzurePolicyDefinitionListArgs;
+pub use azure_policy_definition_show::AzurePolicyDefinitionShowArgs;
 use clap::Args;
 use eyre::Result;